+++ modules/core/src/webgl/accessor.js
@@ -72,7 +72,7 @@ export default class Accessor {
     // TYPE - not expected to be overridden
     if (props.type !== undefined) {
       if (props.type !== this.type) {
-        log.warn('accessor type mismatch');
+        log.warn('accessor type mismatch')();
       }
       this.type = props.type;
 
@@ -85,7 +85,7 @@ export default class Accessor {
     // SIZE - not expected to be overridden
     if (props.size !== undefined) {
       if (this.size !== props.size) {
-        log.warn('accessor size mismatch');
+        log.warn('accessor size mismatch')();
       }
       this.size = props.size;
     }
