+++ test/browser.js
@@ -27,7 +27,7 @@ test.onFinish(() => callExposedFunction('testDone', {success: true}));
 test.onFailure(() => callExposedFunction('testDone', {success: false}));
 
 test('Browser tests', t => {
-  // require('./index-webgl-independent-tests');
+  require('./index-webgl-independent-tests');
   require('./index-webgl-dependent-tests');
   t.end();
 });
