+++ test/modules/core/webgl/copy-and-blit.spec.js
@@ -2,7 +2,7 @@ import test from 'tape-catch';
 import GL from '@luma.gl/constants';
 import {Framebuffer, Renderbuffer, Texture2D, Buffer, getKey} from 'luma.gl';
 import {fixture} from 'luma.gl/test/setup';
-import {TEXTURE_FORMATS} from 'luma.gl/webgl/texture';
+import {TEXTURE_FORMATS} from 'luma.gl/webgl/texture-formats';
 import {
   readPixelsToArray,
   readPixelsToBuffer,
